- extract common use into a prelude mod - extract template filters into own test mod
