+++ tests/integration/helpers/prelude.rs
@@ -1,0 +1,13 @@
+pub use crate::helpers::create_template;
+pub use crate::helpers::project::{binary, Project};
+pub use crate::helpers::project_builder::tempdir;
+
+pub use assert_cmd::prelude::*;
+pub use git2::Repository;
+pub use indoc::indoc;
+pub use predicates::prelude::*;
+pub use std::env;
+pub use std::fs;
+pub use std::ops::Not;
+pub use std::path::PathBuf;
+pub use std::process::Command;
