+++ tests/integration/hooks_and_rhai.rs
@@ -1,13 +1,8 @@
-use indoc::indoc;
-
-use assert_cmd::assert::OutputAssertExt;
-use predicates::str::PredicateStrExt;
-
-use crate::helpers::{project::binary, project_builder::tmp_dir};
+use crate::helpers::prelude::*;
 
 #[test]
 fn it_runs_all_hook_types() {
-    let template = tmp_dir()
+    let template = tempdir()
         .file(
             "init-script.rhai",
             indoc! {r#"
@@ -59,7 +54,7 @@ fn it_runs_all_hook_types() {
         .init_git()
         .build();
 
-    let dir = tmp_dir().build();
+    let dir = tempdir().build();
 
     binary()
         .arg_git(template.path())
@@ -86,7 +81,7 @@ fn it_runs_all_hook_types() {
 
 #[test]
 fn it_runs_system_commands() {
-    let template = tmp_dir()
+    let template = tempdir()
         .file(
             "system-script.rhai",
             indoc! {r#"
@@ -103,7 +98,7 @@ fn it_runs_system_commands() {
         .init_git()
         .build();
 
-    let dir = tmp_dir().build();
+    let dir = tempdir().build();
 
     binary()
         .arg_git(template.path())
@@ -119,7 +114,7 @@ fn it_runs_system_commands() {
 
 #[test]
 fn it_fails_to_prompt_for_system_commands_in_silent_mode() {
-    let template = tmp_dir()
+    let template = tempdir()
         .file(
             "system-script.rhai",
             indoc! {r#"
@@ -136,7 +131,7 @@ fn it_fails_to_prompt_for_system_commands_in_silent_mode() {
         .init_git()
         .build();
 
-    let dir = tmp_dir().build();
+    let dir = tempdir().build();
 
     binary()
         .arg_git(template.path())
@@ -151,7 +146,7 @@ fn it_fails_to_prompt_for_system_commands_in_silent_mode() {
 
 #[test]
 fn it_fails_when_a_system_command_returns_non_zero_exit_code() {
-    let template = tmp_dir()
+    let template = tempdir()
         .file(
             "system-script.rhai",
             indoc! {r#"
@@ -168,7 +163,7 @@ fn it_fails_when_a_system_command_returns_non_zero_exit_code() {
         .init_git()
         .build();
 
-    let dir = tmp_dir().build();
+    let dir = tempdir().build();
 
     binary()
         .arg_git(template.path())
@@ -187,7 +182,7 @@ fn it_fails_when_a_system_command_returns_non_zero_exit_code() {
 
 #[test]
 fn it_fails_when_it_cant_execute_system_command() {
-    let template = tmp_dir()
+    let template = tempdir()
         .file(
             "system-script.rhai",
             indoc! {r#"
@@ -204,7 +199,7 @@ fn it_fails_when_it_cant_execute_system_command() {
         .init_git()
         .build();
 
-    let dir = tmp_dir().build();
+    let dir = tempdir().build();
 
     binary()
         .arg_git(template.path())
@@ -223,7 +218,7 @@ fn it_fails_when_it_cant_execute_system_command() {
 
 #[test]
 fn it_can_change_case() {
-    let template = tmp_dir()
+    let template = tempdir()
         .file(
             "pre-script.rhai",
             indoc! {r#"
@@ -247,7 +242,7 @@ fn it_can_change_case() {
         .init_git()
         .build();
 
-    let dir = tmp_dir().build();
+    let dir = tempdir().build();
 
     binary()
         .arg_git(template.path())
@@ -267,7 +262,7 @@ fn it_can_change_case() {
 
 #[test]
 fn can_change_variables_from_pre_hook() {
-    let template = tmp_dir()
+    let template = tempdir()
         .file(
             "cargo-generate.toml",
             indoc! {r#"
@@ -290,7 +285,7 @@ fn can_change_variables_from_pre_hook() {
         .init_git()
         .build();
 
-    let dir = tmp_dir().build();
+    let dir = tempdir().build();
 
     binary()
         .arg_git(template.path())
@@ -306,7 +301,7 @@ fn can_change_variables_from_pre_hook() {
 
 #[test]
 fn init_hook_can_set_project_name() {
-    let template = tmp_dir()
+    let template = tempdir()
         .file(
             "init.rhai",
             indoc! {r#"
@@ -329,7 +324,7 @@ fn init_hook_can_set_project_name() {
         .init_git()
         .build();
 
-    let dir = tmp_dir().build();
+    let dir = tempdir().build();
 
     binary()
         .arg_git(template.path())
@@ -346,7 +341,7 @@ fn init_hook_can_set_project_name() {
 
 #[test]
 fn init_hook_can_change_project_name_but_keeps_cli_name_for_destination() {
-    let template = tmp_dir()
+    let template = tempdir()
         .file(
             "init.rhai",
             indoc! {r#"
@@ -369,7 +364,7 @@ fn init_hook_can_change_project_name_but_keeps_cli_name_for_destination() {
         .init_git()
         .build();
 
-    let dir = tmp_dir().build();
+    let dir = tempdir().build();
 
     binary()
         .arg_git(template.path())
@@ -385,7 +380,7 @@ fn init_hook_can_change_project_name_but_keeps_cli_name_for_destination() {
 
 #[test]
 fn init_hook_can_change_project_name_but_keeps_init_destination() {
-    let template = tmp_dir()
+    let template = tempdir()
         .file(
             "init.rhai",
             indoc! {r#"
@@ -408,7 +403,7 @@ fn init_hook_can_change_project_name_but_keeps_init_destination() {
         .init_git()
         .build();
 
-    let dir = tmp_dir().build();
+    let dir = tempdir().build();
 
     binary()
         .arg_git(template.path())
@@ -425,7 +420,7 @@ fn init_hook_can_change_project_name_but_keeps_init_destination() {
 
 #[test]
 fn rhai_filter_invokes_rhai_script() {
-    let template = tmp_dir()
+    let template = tempdir()
         .file(
             "filter-script.rhai",
             indoc! {r#"
@@ -441,7 +436,7 @@ fn rhai_filter_invokes_rhai_script() {
         .init_git()
         .build();
 
-    let dir = tmp_dir().build();
+    let dir = tempdir().build();
 
     binary()
         .arg_git(template.path())
@@ -458,7 +453,7 @@ fn rhai_filter_invokes_rhai_script() {
 
 #[test]
 fn missing_rhai_filter_fails_prints_warnings() {
-    let template = tmp_dir()
+    let template = tempdir()
         .file(
             "file_to_expand.txt",
             indoc! {r#"
@@ -468,7 +463,7 @@ fn missing_rhai_filter_fails_prints_warnings() {
         .init_git()
         .build();
 
-    let dir = tmp_dir().build();
+    let dir = tempdir().build();
 
     binary()
         .arg_git(template.path())
