+++ tests/integration/template_config_file/includes_and_excludes.rs
@@ -1,14 +1,8 @@
-use predicates::prelude::*;
-
-use crate::helpers::project::binary;
-use crate::helpers::project_builder::tmp_dir;
-
-use assert_cmd::prelude::*;
-use indoc::indoc;
+use crate::helpers::prelude::*;
 
 #[test]
 fn it_only_processes_include_files_in_config() {
-    let template = tmp_dir()
+    let template = tempdir()
         .file(
             "cargo-generate.toml",
             indoc! {r#"
@@ -23,7 +17,7 @@ fn it_only_processes_include_files_in_config() {
         .init_git()
         .build();
 
-    let dir = tmp_dir().build();
+    let dir = tempdir().build();
 
     binary()
         .arg_git(template.path())
@@ -47,7 +41,7 @@ fn it_only_processes_include_files_in_config() {
 
 #[test]
 fn it_doesnt_process_excluded_files_in_config() {
-    let template = tmp_dir()
+    let template = tempdir()
         .file(
             "cargo-generate.toml",
             indoc! {r#"
@@ -61,7 +55,7 @@ fn it_doesnt_process_excluded_files_in_config() {
         .init_git()
         .build();
 
-    let dir = tmp_dir().build();
+    let dir = tempdir().build();
 
     binary()
         .arg_git(template.path())
@@ -85,7 +79,7 @@ fn it_doesnt_process_excluded_files_in_config() {
 
 #[test]
 fn it_warns_on_include_and_exclude_in_config() {
-    let template = tmp_dir()
+    let template = tempdir()
         .file(
             "Cargo.toml",
             indoc! {r#"
@@ -107,7 +101,7 @@ fn it_warns_on_include_and_exclude_in_config() {
         .init_git()
         .build();
 
-    let dir = tmp_dir().build();
+    let dir = tempdir().build();
 
     binary()
         .arg_git(template.path())
