+++ tests/integration/template_config_file/misc.rs
@@ -1,15 +1,8 @@
-use predicates::prelude::*;
-
-use crate::helpers::project::binary;
-use crate::helpers::project_builder::tmp_dir;
-
-use assert_cmd::prelude::*;
-use git2::Repository;
-use indoc::indoc;
+use crate::helpers::prelude::*;
 
 #[test]
 fn it_always_removes_config_file() {
-    let template = tmp_dir()
+    let template = tempdir()
         .default_manifest()
         .file(
             "cargo-generate.toml",
@@ -19,7 +12,7 @@ fn it_always_removes_config_file() {
         .init_git()
         .build();
 
-    let dir = tmp_dir().build();
+    let dir = tempdir().build();
 
     binary()
         .arg_git(template.path())
@@ -36,7 +29,7 @@ fn it_always_removes_config_file() {
 //https://github.com/ashleygwilliams/cargo-generate/issues/181
 #[test]
 fn it_doesnt_warn_on_config_with_no_ignore() {
-    let template = tmp_dir()
+    let template = tempdir()
         .default_manifest()
         .file(
             "cargo-generate.toml",
@@ -45,7 +38,7 @@ fn it_doesnt_warn_on_config_with_no_ignore() {
         )
         .init_git()
         .build();
-    let dir = tmp_dir().build();
+    let dir = tempdir().build();
 
     binary()
         .arg_git(template.path())
@@ -62,7 +55,7 @@ fn it_doesnt_warn_on_config_with_no_ignore() {
 
 #[test]
 fn a_template_can_specify_to_be_generated_into_cwd() -> anyhow::Result<()> {
-    let template = tmp_dir()
+    let template = tempdir()
         .default_manifest()
         .file(
             "cargo-generate.toml",
@@ -74,7 +67,7 @@ fn a_template_can_specify_to_be_generated_into_cwd() -> anyhow::Result<()> {
         .init_git()
         .build();
 
-    let dir = tmp_dir().build();
+    let dir = tempdir().build();
 
     binary()
         .arg_git(template.path())
@@ -93,7 +86,7 @@ fn a_template_can_specify_to_be_generated_into_cwd() -> anyhow::Result<()> {
 #[test]
 fn vsc_none_can_be_specified_in_the_template() {
     // Build and commit on branch named 'main'
-    let template = tmp_dir()
+    let template = tempdir()
         .default_manifest()
         .file(
             "cargo-generate.toml",
@@ -105,7 +98,7 @@ fn vsc_none_can_be_specified_in_the_template() {
         .init_git()
         .build();
 
-    let dir = tmp_dir().build();
+    let dir = tempdir().build();
 
     binary()
         .arg_git(template.path())
