+++ tests/integration/template_config_file/placeholders.rs
@@ -1,12 +1,8 @@
-use crate::helpers::project::binary;
-use crate::helpers::project_builder::tmp_dir;
-
-use assert_cmd::prelude::*;
-use indoc::indoc;
+use crate::helpers::prelude::*;
 
 #[test]
 fn it_prompts_for_placeholders_in_the_config_file_defined_order() {
-    let template = tmp_dir()
+    let template = tempdir()
         .default_manifest()
         .file(
             "cargo-generate.toml",
@@ -27,7 +23,7 @@ fn it_prompts_for_placeholders_in_the_config_file_defined_order() {
         .init_git()
         .build();
 
-    let dir = tmp_dir().build();
+    let dir = tempdir().build();
 
     binary()
         .arg_git(template.path())
